+++ source/standalone/workflows/robomimic/play.py
@@ -17,6 +17,9 @@ from omni.isaac.orbit.app import AppLauncher
 # add argparse arguments
 parser = argparse.ArgumentParser(description="Play policy trained using robomimic for Orbit environments.")
 parser.add_argument("--cpu", action="store_true", default=False, help="Use CPU pipeline.")
+parser.add_argument(
+    "--disable_fabric", action="store_true", default=False, help="Disable fabric and use USD I/O operations."
+)
 parser.add_argument("--task", type=str, default=None, help="Name of the task.")
 parser.add_argument("--checkpoint", type=str, default=None, help="Pytorch model checkpoint to load.")
 # append AppLauncher cli args
@@ -48,7 +51,7 @@ from omni.isaac.orbit_tasks.utils import parse_env_cfg
 def main():
     """Run a trained policy from robomimic with Orbit environment."""
     # parse configuration
-    env_cfg = parse_env_cfg(args_cli.task, use_gpu=not args_cli.cpu, num_envs=1)
+    env_cfg = parse_env_cfg(args_cli.task, use_gpu=not args_cli.cpu, num_envs=1, use_fabric=not args_cli.disable_fabric)
     # modify configuration
     env_cfg.control.control_type = "inverse_kinematics"
     env_cfg.control.inverse_kinematics.command_type = "pose_rel"